+++ data_linter/validators/great_expectations_validator.py
@@ -1,5 +1,9 @@
 import logging
-from data_linter.validators.base import BaseTableValidator
+from data_linter.validators.base import (
+    BaseTableValidator,
+    ValidatorResult,
+)
+
 from copy import deepcopy
 from typing import List
 
@@ -40,73 +44,6 @@ pd_conversion["character"] = "str"
 log = logging.getLogger("root")
 
 
-class ValidatorResult(object):
-    """
-    Little class to manage adding to validator dict
-    """
-
-    def __init__(self, result_dict={}):
-        if result_dict:
-            self.result = result_dict
-        else:
-            self.result = {"valid": True}
-
-    def get_result(self, copy=True):
-        if copy:
-            return deepcopy(self.result)
-        else:
-            return self.result
-
-    def init_col(self, colname):
-        if colname not in self.result:
-            self.result[colname] = {"valid": True}
-
-    def get_names_of_column_failures(self, test_names: List[str] = []):
-        """
-
-        Return col names which have an overall fail. If test_names is given
-        only returns cols that failed those particular tests is given.
-        Args:
-            test_name (List[str], optional): [description]. List of tests to
-            check against Defaults to [].
-        """
-        non_column_names = ["valid", "validator-table-test-"]
-        failed_cols = []
-        for colname in self.result:
-            if colname in non_column_names:
-                continue
-
-            if test_names:
-                overall_success = True
-                for k, v in self.result[colname].items():
-                    if k in test_names:
-                        overall_success = overall_success and v.get("success", True)
-            else:
-                overall_success = self.result[colname].get("valid", True)
-
-            if not overall_success:
-                failed_cols.append(colname)
-
-        return failed_cols
-
-    def add_table_test(self, testname, test_result):
-        # Same setup - treats overall test as a colname
-        self.init_col(testname)
-        self.result[testname] = test_result
-        if "success" in test_result:
-            self.result["valid"] = self.result["valid"] and test_result["success"]
-
-    def add_test_to_col(self, colname, testname, test_result):
-        self.init_col(colname)
-
-        self.result[colname][testname] = test_result
-        if "success" in test_result:
-            self.result["valid"] = self.result["valid"] and test_result["success"]
-            self.result[colname]["valid"] = (
-                self.result[colname]["valid"] and test_result["success"]
-            )
-
-
 class GreatExpectationsValidator(BaseTableValidator):
     """
     Great expectations data validator
@@ -132,7 +69,6 @@ class GreatExpectationsValidator(BaseTableValidator):
         self.default_result_fmt = default_result_fmt
         self.ignore_missing_cols = ignore_missing_cols
 
-        self.response = ValidatorResult()
         self.valid = self.response.result["valid"]
 
     def write_validation_errors_to_log(self):
